requirements.txt
pynput==1.7.6
pillow==10.2.0